UML-33 refactoring resizer
