+++ package.json
@@ -41,6 +41,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-transition-group": "^4.4.2",
+    "reset-css": "^5.0.1",
     "resize-observer-polyfill": "^1.5.1",
     "tplant": "^3.1.0",
     "typescript": "^4.3.5"
