--- src/components/DiagramEditor/DiagramEditor.module.scss
@@ -1,3 +1,0 @@
-.container {
-  height: 100%;
-}
