+++ src/components/SplitPane/Pane/Pane.module.scss
@@ -1,4 +1,5 @@
 .pane {
   width: 50%;
+  height: 100%;
   z-index: 0;
 }
