+++ src/components/SplitPane/SplitPane.module.scss
@@ -1,3 +1,5 @@
 .SplitPane {
+  height: 100%;
   display: flex !important;
+  justify-content: space-between;
 }
