+++ src/styles/index.scss
@@ -1,1 +1,2 @@
+@import 'node_modules/reset-css/sass/reset';
 @import "main";
